map ListOffers

+++ main.py
@@ -105,8 +105,8 @@ def mapPage():
         listOffers=[]
         for obj in results['results']['bindings']:
             listOffers.append([obj['label_100']['value'],
-                double(obj['label_long']['value']),
-                double(obj['label_lat']['value'])])
+                float(obj['label_long']['value']),
+                float(obj['label_lat']['value'])])
 
         print(listOffers)
